Simplify test suite
