+++ .gitignore
@@ -14,6 +14,7 @@
 /test/dummy/.bundle/config
 /test/dummy/rails_pulse_test*
 /test/dummy/db/schema.rb
+test/dummy/rails_pulse_development
 
 # Environment variables
 .env
