+++ test/test_helper.rb
@@ -13,11 +13,10 @@ end
 
 require_relative "../test/dummy/config/environment"
 
-# Force database configuration early for MySQL in CI
-if ENV["DB"] == "mysql2" || ENV["DB"] == "mysql"
+# Configure database early if needed
+if ENV["FORCE_DB_CONFIG"] == "true"
   require_relative "support/database_helpers"
   DatabaseHelpers.configure_test_database
-
 end
 # Only use dummy app migrations to avoid conflicts
 ActiveRecord::Migrator.migrations_paths = [
@@ -41,43 +40,15 @@ rescue LoadError
 end
 
 # Load test dependencies
-begin
-  require "mocha/minitest"
-rescue LoadError
-  puts "Warning: mocha not available for testing"
-end
+require "mocha/minitest"
+require "timecop"
+require "factory_bot_rails"
+require "shoulda-matchers"
+require "minitest/reporters"
+require "database_cleaner/active_record"
 
-begin
-  require "timecop"
-rescue LoadError
-  puts "Warning: timecop not available for testing"
-end
-
-begin
-  require "factory_bot_rails"
-rescue LoadError
-  puts "Warning: factory_bot_rails not available for testing"
-end
-
-begin
-  require "shoulda-matchers"
-rescue LoadError
-  puts "Warning: shoulda-matchers not available for testing"
-end
-
-begin
-  require "minitest/reporters"
-  Minitest::Reporters.use! Minitest::Reporters::ProgressReporter.new
-  # Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
-rescue LoadError
-  puts "Warning: minitest-reporters not available for testing"
-end
-
-begin
-  require "database_cleaner/active_record"
-rescue LoadError
-  puts "Warning: database_cleaner not available for testing"
-end
+# Use progress reporter for cleaner output
+Minitest::Reporters.use! Minitest::Reporters::ProgressReporter.new
 
 # Load support files
 Dir[File.expand_path("support/**/*.rb", __dir__)].each { |f| require f }
@@ -90,11 +61,9 @@ class ActiveSupport::TestCase
   # Include test helpers
   include DatabaseHelpers
   include PerformanceHelpers
-  include StubHelpers
   include ModelTestHelpers
   include ControllerTestHelpers
   include FactoryHelpers
-  include PerformanceTestHelpers
 
   # Configure FactoryBot
   if defined?(FactoryBot)
@@ -123,24 +92,12 @@ class ActiveSupport::TestCase
     DatabaseCleaner.start
   end
 
-  # Configure database for speed
-  def self.configure_database_for_test_type
-    test_type = ENV.fetch("TEST_TYPE", "unit")
-
-    case test_type
-    when "unit", "functional"
-      # Use memory database for unit/functional tests (fastest)
-      ENV["MEMORY_DATABASE"] = "true"
-    when "integration", "system"
-      # Use file database for integration tests (more realistic)
-      ENV["MEMORY_DATABASE"] = "false"
-    end
-
-    DatabaseHelpers.configure_test_database
-  end
 
   # Fast setup/teardown using transactions
   setup do
+    # Ensure Rails Pulse tables exist for this test
+    DatabaseHelpers.ensure_test_tables_exist
+
     setup_test_database if respond_to?(:setup_test_database)
 
     # Stub time operations only
@@ -166,8 +123,6 @@ class ActiveSupport::TestCase
   end
 end
 
-# Configure database based on test type
-ActiveSupport::TestCase.configure_database_for_test_type
 
 # Database switching based on environment variables (must happen after Rails initialization)
 if ENV["FORCE_DB_CONFIG"] == "true"
@@ -182,43 +137,17 @@ end
 # Always ensure Rails Pulse tables exist BEFORE any tests run
 DatabaseHelpers.ensure_test_tables_exist
 
-# Force table creation immediately in CI to avoid timing issues
-if ENV["CI"] == "true"
-  puts "Forcing table verification in CI..."
-  required_tables = [ "rails_pulse_routes", "rails_pulse_requests", "rails_pulse_queries", "rails_pulse_operations", "rails_pulse_summaries" ]
-
-  # Use the same connection that Rails Pulse models will use
-  connection = defined?(RailsPulse::ApplicationRecord) ? RailsPulse::ApplicationRecord.connection : ActiveRecord::Base.connection
-  missing_tables = required_tables.reject { |table| connection.table_exists?(table) }
-
-  if missing_tables.any?
-    puts "FATAL: Required tables missing after creation: #{missing_tables.join(', ')}"
-    puts "Available tables: #{connection.tables.sort.join(', ')}"
-    puts "Database: #{connection.connection_db_config.database}"
-    puts "Adapter: #{connection.adapter_name}"
-    exit 1
-  end
-  puts "All required tables confirmed present."
-end
 
 # Display test environment information
-puts "\n" + "=" * 80
-puts "🚀 Rails Pulse Test Suite"
-puts "=" * 80
-puts "Ruby version:       #{RUBY_VERSION}"
-puts "Rails version:      #{Rails.version}"
-requested_adapter = ENV["DB"] || "sqlite3"
-display_adapter = requested_adapter == "sqlite3" && ENV["FORCE_DB_CONFIG"] != "true" ? "sqlite3 (default)" : requested_adapter
-actual_adapter = ActiveRecord::Base.connection.adapter_name
-puts "DB:                 #{display_adapter}"
-puts "Actual adapter:     #{actual_adapter}"
-if ENV["FORCE_DB_CONFIG"] != "true" && requested_adapter != "sqlite3"
-  puts "📝 Note: Database switching disabled by default. Set FORCE_DB_CONFIG=true to enable."
+if ENV["VERBOSE"] == "true"
+  puts "\n" + "=" * 50
+  puts "🚀 Rails Pulse Test Suite"
+  puts "=" * 50
+  puts "Rails version: #{Rails.version}"
+  puts "Database: #{ENV['DB'] || 'sqlite3'}"
+  puts "=" * 50
+  puts
 end
-puts "Database name:      #{ActiveRecord::Base.connection_db_config.database}"
-puts "Test environment:   #{Rails.env}"
-puts "=" * 80
-puts
 
 # Load fixtures from the engine (temporarily disabled to avoid foreign key issues)
 if ActiveSupport::TestCase.respond_to?(:fixture_paths=)
